Add a filter for backend selector
A filter for backend selector has been added.
In order to avoid a very large list of providers,
this filter can be used to only display the
transportation in a given country.

API for backend desktop files has been changed
a bit as well, since countries should be provided
using a country code.